Support @deprecated tag in DocsGen
