+++ uui-build/ts/tasks/docsGen/tsdoc/tsDocUtils.ts
@@ -1,14 +1,12 @@
 import {
-    DocBlock,
     DocComment,
-    DocPlainText,
     ParserContext,
-    Standardization,
     TSDocConfiguration,
     TSDocParser,
-    TSDocTagSyntaxKind,
 } from '@microsoft/tsdoc';
 import { TComment } from '../types/sharedTypes';
+import { cleanAsteriks } from './tsDocFormatters';
+import { CORE_TAGS, CUSTOM_TAGS } from './tsDocTags';
 
 export class TsDocUtils {
     static isCommentEmpty(comment: TComment | undefined): boolean {
@@ -54,7 +52,7 @@ export class TsDocUtils {
                 if (parserRes) {
                     const { extracted, value } = parserRes;
                     if (extracted) {
-                        acc[name] = value;
+                        acc[name] = value as any;
                     }
                 } else {
                     console.error(`Unable to parse tag=${name} from comment below. So, this tag will be ignored. Comment:\n${rawComment}`);
@@ -62,65 +60,19 @@ export class TsDocUtils {
             }
             return acc;
         }, {});
+
+        if (docComment.deprecatedBlock) {
+            const parserRes = CORE_TAGS['@deprecated'].parser(docComment.deprecatedBlock);
+            if (parserRes) {
+                const { extracted, value } = parserRes;
+                if (extracted) {
+                    tags!['@deprecated'] = value;
+                }
+            }
+        }
+
         if (Object.keys(tags || {}).length > 0) {
             return tags;
         }
     }
-}
-
-const CUSTOM_TAGS = {
-    '@default': {
-        tagDefinition: {
-            tagName: '@default',
-            syntaxKind: TSDocTagSyntaxKind.BlockTag,
-            standardization: Standardization.Discretionary,
-            allowMultiple: false,
-            tagNameWithUpperCase: '@DEFAULT',
-        },
-        parser: (db: DocBlock) => {
-            const valueParts = db.getChildNodes()[1]; // DocSection
-            const firstValuePart = valueParts.getChildNodes()[0]; // DocParagraph
-            const txt = firstValuePart.getChildNodes()[0]; // DocPlainText
-            const isInQuotes = (t: string) => ["'", '"'].indexOf(t[0]) !== -1 && ["'", '"'].indexOf(t[t.length - 1]) !== -1;
-            if (txt instanceof DocPlainText) {
-                const txtValue = txt?.text?.trim();
-                if (typeof txtValue !== 'undefined') {
-                    let value: boolean | string | number | null;
-
-                    switch (txtValue) {
-                        case 'true':
-                        case 'false': {
-                            value = txtValue === 'true';
-                            break;
-                        }
-                        case 'null': {
-                            value = null;
-                            break;
-                        }
-                        default: {
-                            if (!isNaN(+txtValue)) {
-                                value = +txtValue;
-                            } else if (isInQuotes(txtValue)) {
-                                value = txtValue.substring(1, txtValue.length - 1);
-                            } else {
-                                return;
-                            }
-                            break;
-                        }
-                    }
-                    return {
-                        extracted: true,
-                        value,
-                    };
-                }
-            }
-        },
-    },
-};
-
-function cleanAsteriks(line: string): string {
-    const regex1 = /^([\s]*\/[*]{1,2})(.*)$/; // leading /* or /**
-    const regex2 = /^(.*)([\s]*[*]{1,2}\/)$/; // trailing */ or **/
-    const regex3 = /^([\s]*[*]{1,1})(.*)$/; // leading *
-    return line.replace(regex1, '$2').replace(regex2, '$1').replace(regex3, '$2');
 }
